+++ launchDaemonManagement.zsh
@@ -21,13 +21,12 @@
 #
 # HISTORY
 #
-# Version 2.2.0b11, 17-Dec-2025, Dan K. Snelson (@dan-snelson)
-#   - Addressed Feature Request: Intelligently display reminder dialog after rebooting #42
-#   - Added instructions for monitoring the client-side log
-#   - `assemble.zsh` now outputs to `Artifacts/` (instead of `Resources/`)
-#   - Updated `Resources/sample.plist` to address Feature Request #43
-#   - Harmonized Organization Variables between `launchDaemonManagement.zsh` and `reminderDialog.zsh`
-#   - Added Detection for staged macOS updates (Addresses Feature Request #49)
+# Version 2.2.0b12, 17-Dec-2025, Dan K. Snelson (@dan-snelson)
+# - Added "quiet period" to skip reminder dialog if recently shown (Addresses Feature Request #42)
+# - Added instructions for monitoring the client-side log to the log file itself
+# - `assemble.zsh` now outputs to `Artifacts/` (instead of `Resources/`)
+# - Updated `Resources/sample.plist` to address Feature Request #43
+# - Added Detection for staged macOS updates (Addresses Feature Request #49)
 #
 ####################################################################################################
 
@@ -42,7 +41,7 @@
 export PATH=/usr/bin:/bin:/usr/sbin:/sbin:/usr/local:/usr/local/bin
 
 # Script Version
-scriptVersion="2.2.0b11"
+scriptVersion="2.2.0b12"
 
 # Client-side Log
 scriptLog="/var/log/org.churchofjesuschrist.log"
@@ -136,7 +135,7 @@ function resetConfiguration() {
             # Reset LaunchDaemon
             info "Reset LaunchDaemon … "
             launchDaemonStatus
-            if [[ -n "${launchDaemonStatus}" ]]; then
+            if [[ -n "${launchDaemonStatusResult}" ]]; then
                 logComment "Unload '${launchDaemonPath}' … "
                 launchctl bootout system "${launchDaemonPath}"
                 launchDaemonStatus
@@ -156,7 +155,7 @@ function resetConfiguration() {
 
             info "Reset LaunchDaemon … "
             launchDaemonStatus
-            if [[ -n "${launchDaemonStatus}" ]]; then
+            if [[ -n "${launchDaemonStatusResult}" ]]; then
                 logComment "Unload '${launchDaemonPath}' … "
                 launchctl bootout system "${launchDaemonPath}"
                 launchDaemonStatus
@@ -181,7 +180,7 @@ function resetConfiguration() {
             # Uninstall LaunchDaemon
             info "Uninstall LaunchDaemon … "
             launchDaemonStatus
-            if [[ -n "${launchDaemonStatus}" ]]; then
+            if [[ -n "${launchDaemonStatusResult}" ]]; then
                 logComment "Unload '${launchDaemonPath}' … "
                 launchctl bootout system "${launchDaemonPath}"
                 launchDaemonStatus
@@ -361,10 +360,10 @@ function launchDaemonStatus() {
 
     notice "LaunchDaemon Status"
     
-    launchDaemonStatus=$( launchctl list | grep "${launchDaemonLabel}" )
+    launchDaemonStatusResult=$( launchctl list | grep "${launchDaemonLabel}" )
 
-    if [[ -n "${launchDaemonStatus}" ]]; then
-        logComment "${launchDaemonStatus}"
+    if [[ -n "${launchDaemonStatusResult}" ]]; then
+        logComment "${launchDaemonStatusResult}"
     else
         logComment "${launchDaemonLabel} is NOT loaded"
     fi
@@ -552,14 +551,14 @@ if [[ -f "${launchDaemonPath}" ]]; then
 
     launchDaemonStatus
 
-    if [[ -n "${launchDaemonStatus}" ]]; then
+    if [[ -n "${launchDaemonStatusResult}" ]]; then
 
         logComment "${launchDaemonLabel} IS loaded"
 
     else
 
         logComment "Loading '${launchDaemonLabel}' …"
-        launchctl asuser $(id -u) bootstrap gui/$(id -u) "${launchDaemonPath}"
+        launchctl bootstrap system "${launchDaemonPath}"
         launchDaemonStatus
 
     fi